# improve performance of getAllReportErrors