--- a/src/libs/OptionsListUtils.ts
+++ b/src/libs/OptionsListUtils.ts
@@ -496,6 +496,7 @@ function getAllReportErrors(report: OnyxEntry<Report>, reportActions: OnyxEntry<
         reportActionErrors.smartscan = ErrorUtils.getMicroSecondOnyxErrorWithTranslationKey('iou.error.genericSmartscanFailureMessage');
     }
     // All error objects related to the report. Each object in the sources contains error messages keyed by microtime
+    // Use Object.assign to merge all error objects into one since it is faster and uses less memory than spread operator
     // eslint-disable-next-line prefer-object-spread
     const errorSources = Object.assign({}, reportErrors, reportErrorFields, reportActionErrors);